refactor(experience): move logo utils to shared folder
Move getBrandingLogoUrl and getLogoUrl to shared/utils for account package access
